+++ arviz/plots/backends/bokeh/parallelplot.py
@@ -46,13 +46,21 @@ def plot_parallel(
     figsize, *_ = _scale_fig_size(figsize, textsize, 1, 1)
 
     if ax is None:
-        ax = create_axes_grid(1, figsize=figsize, squeeze=True, backend_kwargs=backend_kwargs,)
+        ax = create_axes_grid(
+            1,
+            figsize=figsize,
+            squeeze=True,
+            backend_kwargs=backend_kwargs,
+        )
 
     non_div = list(posterior[:, ~diverging_mask].T)
     x_non_div = [list(range(len(non_div[0]))) for _ in range(len(non_div))]
 
     ax.multi_line(
-        x_non_div, non_div, line_color="black", line_alpha=0.05,
+        x_non_div,
+        non_div,
+        line_color="black",
+        line_alpha=0.05,
     )
 
     if np.any(diverging_mask):